+++ settei/base.py
@@ -21,7 +21,6 @@ __all__ = ('ConfigError', 'ConfigKeyError', 'ConfigTypeError',
            'Configuration', 'ConfigValueError', 'ConfigWarning',
            'config_object_property', 'config_property', 'get_union_types')
 
-
 if hasattr(typing, 'UnionMeta'):
     def get_union_types(type_) -> bool:
         """Return a :class:`tuple` of the given :class:`~typing.Union`
@@ -47,7 +46,7 @@ else:
     # For newer versions of typing (>= Python 3.7)
     def get_union_types(type_) -> bool:
         if getattr(type_, '__origin__', None) is typing.Union and \
-           hasattr(type_, '__args__'):
+                hasattr(type_, '__args__'):
             return type_.__args__
 
 
@@ -79,6 +78,10 @@ class config_property:
                             this option is only available when ``default``
                             value is provided
     :type default_warning: :class:`bool`
+    :type cached: :class:`bool`
+    :param cached: keyword only argument.
+                   get config value which is cached on its instance so that
+                   config value won't be created again.
 
     .. versionchanged:: 0.4.0
 
@@ -94,9 +97,16 @@ class config_property:
     """
 
     @typechecked
-    def __init__(self, key: str, cls, docstring: str = None, **kwargs) -> None:
+<<<<<<< HEAD
+    def __init__(self, key: str, cls, docstring: str=None, cached: bool=False,
+=======
+    def __init__(self, key: str, cls, docstring: str = None,
+                 *, cached: bool = False,
+>>>>>>> 9729869... lint check - flake8
+                 **kwargs) -> None:
         self.key = key
         self.cls = cls
+        self.cached = cached
         self.__doc__ = docstring
         if 'default_func' in kwargs:
             if 'default' in kwargs:
@@ -121,13 +131,26 @@ class config_property:
             self.default_func = None
             self.default_warning = False
 
-    def __get__(self, obj, cls: typing.Optional[type] = None):
+    def raw_value(self, obj):
         if obj is None:
             return self
         default, value = self.get_raw_value(obj)
         if not default:
             value = self.convert_native_type(value)
             self.typecheck(value)
+        return value
+
+    def __get__(self, obj, cls: typing.Optional[type] = None):
+        if self.cached:
+            cache_key = '  cache_{!s}'.format(self.key)
+            instance = getattr(obj, cache_key, None)
+            if instance is None:
+                value = self.raw_value(obj)
+                setattr(obj, cache_key, value)
+            else:
+                value = instance
+        else:
+            value = self.raw_value(obj)
         return value
 
     def get_raw_value(self, obj) -> typing.Tuple[bool, object]:
@@ -356,7 +379,7 @@ class config_object_property(config_property):
         f = self.import_(import_path)
         args = expression.get('*', ())
         if isinstance(args, str) or \
-           not isinstance(args, collections.abc.Sequence):
+                not isinstance(args, collections.abc.Sequence):
             raise ConfigValueError(
                 '"*" field must be a list, not ' + repr(args)
             )